pyproject.toml
[tool.poetry]
name = "telegram-audio-extractor-bot"
version = "0.1.0"
description = "Async Telegram bot to extract audio from video links using ffmpeg."
authors = ["You <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.7.0"
aiohttp = "~3.9.5"
python-dotenv = "^1.0.1"
pydantic = "^2.7.0, <2.8.0"
yt-dlp = "^2024.8.6"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "^0.23.8"
pytest-mock = "^3.14.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
